Reorganize (#3)
* Reorganized code.

* cleanup dockerfile
