+++ .github/workflows/bandit.yml
@@ -28,4 +28,4 @@ jobs:
         # Version range or exact version of a Python version to use, using SemVer's version range syntax.
         python_version:  3.7.1
         # This is the path at which the your python project is located in you github repo.
-        project_path: annotation.py
+        project_path: main.py
