--- CONTRIBUTING.md
@@ -1,19 +1,0 @@
-## Contributing
-
-[code-of-conduct]: CODE_OF_CONDUCT.md
-
-Hi there 👋 We are excited that you want to contribute a new workflow to this repo. By doing this you are helping people get up and running with GitHub Actions and that's cool 😎.
-
-Contributions to this project are [released](https://help.github.com/articles/github-terms-of-service/#6-contributions-under-repository-license) to the public under the [project's open source license](https://github.com/Joel-hanson/bandit-report-artifacts/blob/master/LICENSE).
-
-Please note that this project is released with a [Contributor Code of Conduct](
-https://github.com/Joel-hanson/bandit-report-artifacts/blob/master/.github/CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
-
-Before merging a new workflow, the following requirements need to be met:
-
-- Should be as simple as is needed for the service.
-- There are many programming languages and tools out there. Right now we don't have a page that allows for a really large number of workflows, so we do have to be a little choosy about what we accept. Less popular tools or languages might not be accepted.
-- Should not send data to any 3rd party service except for the purposes of installing dependencies. If so needed please notify the concerned.
-- Cannot be to a paid service or product.
-
-Thank you
