+++ Dockerfile
@@ -2,16 +2,11 @@ FROM python:3.8-slim-buster
 
 RUN apt-get update
 
-# RUN apt-get install -y make build-essential libssl-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev wget curl llvm libncurses5-dev git libpq-dev
-RUN apt-get install -y curl git
-
-RUN curl -L https://raw.githubusercontent.com/yyuu/pyenv-installer/master/bin/pyenv-installer | bash
-
-ENV PYENV_ROOT /root/.pyenv
-ENV PATH $PYENV_ROOT/shims:$PYENV_ROOT/bin:$PATH
-
+ADD entrypoint.sh /
 ADD requirements.txt /
-ADD annotation.py /
-ADD entrypoint.sh /entrypoint.sh
+ADD main.py /
+
+RUN pip install -r /requirements.txt
 RUN chmod +x /entrypoint.sh
+
 ENTRYPOINT ["/entrypoint.sh"]
